// frontend/src/api/models/ClientCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClientCreate = {
    full_name: string;
    phone: string;
    email?: (string | null);
    city?: (string | null);
    area?: (string | null);
    relation?: (string | null);
};