// frontend/src/api/models/ClientUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClientUpdate = {
    full_name?: (string | null);
    phone?: (string | null);
    email?: (string | null);
    city?: (string | null);
    area?: (string | null);
    relation?: (string | null);
};